move cron logic back to common
